refactor: centralize document handling
